fix: asset image badges on my assets -> land
